requirements.txt
lxml==3.4.4
requests==2.7.0
unicodecsv==0.14.1
wheel==0.24.0
PyMySQL == 0.6.2
Pillow == 3.2.0
pyserial == 3.0.1